requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
python-dotenv>=1.0.0
pygithub>=1.59.0
pyarrow>=12.0.0
fastparquet>=0.8.3
plotly>=5.0.0